added HUD to show ammo and mag count, added physics material to player to stop from sticking to walls

+++ Assets/Scripts/WeaponPickupObject.cs
@@ -6,6 +6,6 @@ public class WeaponPickupObject : MonoBehaviour
 {
 
     public GameObject gunObject;
-    public float currentAmmo;
+    public int currentAmmo;
 
 }
